+++ models/menu.py
@@ -14,15 +14,8 @@ class Menu:
         """
         Inicializa la estructura de datos para almacenar productos
         """
-        # self.productos = {}
-        "Verifica la existencia de la tabla"
         self.bd = BD('happyburger.db')
-        # if self.bd.verificarBaseDatosExiste():
-        #     pass
-        # else:
-        #     self.bd.crearBaseDatos()
-        #     self.bd.crearTablaMenu()
-    
+
     def mostrarListaMenu(self):
         """
         Regresa una lista de productos registrados
@@ -44,7 +37,6 @@ class Menu:
         except sqlite3.Error as e:
             print("Error al intentar obtener el menú: {}".format(e))
         finally:
-            print("--------------------------------------------------")
             if conexion:
                 cursor.close()
                 conexion.close()
@@ -71,7 +63,32 @@ class Menu:
         except Exception as e:
             print("Error al buscar el producto: {}".format(e))
         finally:
-            print("--------------------------------------------------")
+            if conexion:
+                cursor.close()
+                conexion.close()
+    
+    def buscarProductoPorNombre(self, nombre_producto):
+        """
+        Busca y regresa un producto por su nombre
+        """
+        try: 
+            conexion = self.bd.abrirConexion()
+            cursor = conexion.cursor()
+            cursor.execute("SELECT * FROM menu WHERE nombre = ?", (nombre_producto,))
+            producto = cursor.fetchone()
+            if not producto:
+                print(f"No se encontró un producto con el nombre {nombre_producto}")
+                return producto
+
+            return {
+                "id": producto[0],
+                "clave": producto[1],
+                "nombre": producto[2],
+                "precio": producto[3]
+            }
+        except Exception as e:
+            print("Error al buscar el producto: {}".format(e))
+        finally:
             if conexion:
                 cursor.close()
                 conexion.close()
@@ -81,11 +98,15 @@ class Menu:
         Agrega un nuevo producto al menú
         """
         try:
+            conexion = self.bd.abrirConexion()
+            cursor = conexion.cursor()
             print("--------------------------------------------------")
             print("Nuevo producto")
             producto = self.ingresarDatosProducto()
-            conexion = self.bd.abrirConexion()
-            cursor = conexion.cursor()
+            if not producto:
+                print("Operación cancelada")
+                return
+            
             sql = '''INSERT INTO menu(clave, nombre, precio)
             VALUES(?,?,?)'''
             cursor.execute(sql, (producto["clave"], producto["nombre"], producto["precio"],))
@@ -109,10 +130,19 @@ class Menu:
             conexion = self.bd.abrirConexion()
             cursor = conexion.cursor()
             self.mostrarListaMenu()
-            id_menu = Utils.ingresarId("Ingrese el id del producto a editar: ")
+            print("--------------------------------------------------")
+            id_menu = Utils.ingresarId("Ingrese el id del producto a editar (para cancelar teclee la palabra esc): ")
+            if not id_menu:
+                print("Operación cancelada")
+                return
+            
             if not self.buscarProducto(id_menu):
                 return
             producto = self.ingresarDatosProducto()
+            if not producto:
+                print("Operación cancelada")
+                return
+            
             sql ="UPDATE menu SET clave = ?, nombre = ?, precio = ? WHERE id = ?"
             valores =  (producto["clave"], producto["nombre"], producto["precio"], id_menu)
             cursor.execute(sql, valores)
@@ -123,7 +153,6 @@ class Menu:
         except Exception as e:
             print("Error:".format(e))
         finally:
-            print("--------------------------------------------------")
             if conexion:
                 cursor.close()
                 conexion.close()
@@ -136,7 +165,12 @@ class Menu:
             conexion = self.bd.abrirConexion()
             cursor = conexion.cursor()
             self.mostrarListaMenu()
-            id_menu = Utils.ingresarId("Ingrese el id del producto a eliminar: ")
+            print("--------------------------------------------------")
+            id_menu = Utils.ingresarId("Ingrese el id del producto a eliminar (para cancelar teclee la palabra esc): ")
+            if not id_menu:
+                print("Operación cancelada")
+                return
+            
             if not self.buscarProducto(id_menu):
                 return
             sql = '''DELETE FROM menu WHERE id = ?'''
@@ -157,9 +191,14 @@ class Menu:
         datosIncorrectos = True
         while datosIncorrectos:
             try:
+                print("Datos del producto (para cancelar teclee la palabra esc)")
                 clave = input("Ingrese la clave del producto: ")
+                if clave == "esc": return {}
                 nombre = input("Ingrese en nombre del producto: ")
-                precio = float(input("Ingrese el precio del producto: "))
+                if nombre == "esc": return {}
+                precio = input("Ingrese el precio del producto: ")
+                if precio == "esc": return {}
+                precio = float(precio)
                 datosIncorrectos = False
             except Exception as e:
                 print('Error al capturar un dato: {}'.format(e))
@@ -169,22 +208,3 @@ class Menu:
             "nombre": nombre,
             "precio": precio
         }
-    
-    # def ingresarId(self, mensaje):
-    #     """
-    #     Solicita y retorna el id de un producto
-    #     """
-    #     id_menu = 0
-    #     id_incorrecto = True
-    #     while id_incorrecto:
-    #         try:
-    #             id_menu = int(input(mensaje))
-    #             id_incorrecto = False
-    #         except Exception as e:
-    #             print("Error al capturar el id: {}".format(e))
-    #             print("Intente de nuevo ingresar el id \n")
-    #     return id_menu
-  
-
-
-
